+++ src/lib/y2network/dialogs/edit_interface.rb
@@ -20,8 +20,8 @@
 require "cwm/dialog"
 
 require "y2network/widgets/address_tab.rb"
-require "y2network/widgets/bond_slaves_tab.rb"
-require "y2network/widgets/bridge_slaves_tab.rb"
+require "y2network/widgets/bond_ports_tab.rb"
+require "y2network/widgets/bridge_ports_tab.rb"
 require "y2network/widgets/general_tab.rb"
 require "y2network/widgets/hardware_tab.rb"
 require "y2network/widgets/wireless_tab.rb"
