+++ src/lib/y2network/widgets/bridge_ports_tab.rb
